[4.6, 4.8] Add -mtune=intel support
This is backport of r205719 and r205754 from trunk.

Now -mtune=intel matches -mtune=slm for 4.8 and -mtune=atom for 4.6.

As written in gcc-4.8 docs:

Produce code optimized for the most current Intel processors, which are
Haswell and Silvermont for this version of GCC.  If you know the CPU
on which your code will run, then you should use the corresponding
-mtune or -march option instead of -mtune=intel.
But, if you want your application performs better on both Haswell and
Silvermont, then you should use this option.

As new Intel processors are deployed in the marketplace, the behavior of
this option will change.  Therefore, if you upgrade to a newer version of
GCC, code generation controlled by this option will change to reflect
the most current Intel processors at the time that version of GCC is
released.

There is no -march=intel option because -march indicates
the instruction set the compiler can use, and there is no common
instruction set applicable to all processors.  In contrast,
-mtune indicates the processor (or, in this case, collection of
processors) for which the code is optimized.

Change-Id: I3ec4c3b5423d9b3547cd8e3aa77a18af3fd89598
Signed-off-by: Pavel Chupin <pavel.v.chupin@intel.com>