Only package gccunwind for Linux.
We should be doing this for the others, but there are still some build
issues to work out.

Note that this change was made to ndk/, but that was done after I had
copied build-gcc.sh here and I mistakenly forgot to apply it here as
well.

Change-Id: I2289c9d4da41eee8ce1d75324ccabf30adb7bd01
(cherry picked from commit da6ed85636572e47535ffd4fe6ab4790921c6508)